# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice Translation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface InitiateCallRequest {
  /** Phone number to call in E.164 format */
  toNumber: string;
  /** Language code of the caller (e.g. en, es, fr) */
  sourceLanguage: string;
  /** Language code to translate to (e.g. es, en, zh) */
  targetLanguage: string;
}

export interface InitiateCallResponse {
  callSid: string;
  status: string;
  message?: string;
}

export interface TranscriptEntry {
  id: string;
  timestamp: string;
  originalText: string;
  translatedText: string;
  speaker: string;
}

export interface CallStatus {
  callSid: string;
  status: string;
  sourceLanguage?: string;
  targetLanguage?: string;
  duration?: number;
  transcript?: TranscriptEntry[];
}

export interface CallRecord {
  callSid: string;
  toNumber: string;
  sourceLanguage: string;
  targetLanguage: string;
  startedAt: string;
  endedAt?: string | null;
  duration?: number | null;
  status: string;
  transcriptCount: number;
}

export interface Language {
  code: string;
  name: string;
  nativeName: string;
}

export interface TranslateTextRequest {
  text: string;
  sourceLanguage: string;
  targetLanguage: string;
}

export interface TranslateTextResponse {
  originalText: string;
  translatedText: string;
  sourceLanguage: string;
  targetLanguage: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}