# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice Translation API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Initiate an outbound translated call
 */
export const InitiateCallBody = zod.object({
  toNumber: zod.string().describe("Phone number to call in E.164 format"),
  sourceLanguage: zod
    .string()
    .describe("Language code of the caller (e.g. en, es, fr)"),
  targetLanguage: zod
    .string()
    .describe("Language code to translate to (e.g. es, en, zh)"),
});

export const InitiateCallResponse = zod.object({
  callSid: zod.string(),
  status: zod.string(),
  message: zod.string().optional(),
});

/**
 * @summary Get status of an active call
 */
export const GetCallStatusParams = zod.object({
  callSid: zod.coerce.string(),
});

export const GetCallStatusResponse = zod.object({
  callSid: zod.string(),
  status: zod.string(),
  sourceLanguage: zod.string().optional(),
  targetLanguage: zod.string().optional(),
  duration: zod.number().optional(),
  transcript: zod
    .array(
      zod.object({
        id: zod.string(),
        timestamp: zod.string(),
        originalText: zod.string(),
        translatedText: zod.string(),
        speaker: zod.string(),
      }),
    )
    .optional(),
});

/**
 * @summary End an active call
 */
export const EndCallParams = zod.object({
  callSid: zod.coerce.string(),
});

export const EndCallResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Get recent call history with translations
 */
export const GetCallHistoryResponseItem = zod.object({
  callSid: zod.string(),
  toNumber: zod.string(),
  sourceLanguage: zod.string(),
  targetLanguage: zod.string(),
  startedAt: zod.string(),
  endedAt: zod.string().nullish(),
  duration: zod.number().nullish(),
  status: zod.string(),
  transcriptCount: zod.number(),
});
export const GetCallHistoryResponse = zod.array(GetCallHistoryResponseItem);

/**
 * @summary Get list of supported languages
 */
export const GetSupportedLanguagesResponseItem = zod.object({
  code: zod.string(),
  name: zod.string(),
  nativeName: zod.string(),
});
export const GetSupportedLanguagesResponse = zod.array(
  GetSupportedLanguagesResponseItem,
);

/**
 * @summary Translate text (test endpoint)
 */
export const TranslateTextBody = zod.object({
  text: zod.string(),
  sourceLanguage: zod.string(),
  targetLanguage: zod.string(),
});

export const TranslateTextResponse = zod.object({
  originalText: zod.string(),
  translatedText: zod.string(),
  sourceLanguage: zod.string(),
  targetLanguage: zod.string(),
});