# lib/api-zod/src/generated/types/callRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice Translation API
 * OpenAPI spec version: 0.1.0
 */

export interface CallRecord {
  callSid: string;
  toNumber: string;
  sourceLanguage: string;
  targetLanguage: string;
  startedAt: string;
  endedAt?: string | null;
  duration?: number | null;
  status: string;
  transcriptCount: number;
}