# lib/api-zod/src/generated/types/initiateCallRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice Translation API
 * OpenAPI spec version: 0.1.0
 */

export interface InitiateCallRequest {
  /** Phone number to call in E.164 format */
  toNumber: string;
  /** Language code of the caller (e.g. en, es, fr) */
  sourceLanguage: string;
  /** Language code to translate to (e.g. es, en, zh) */
  targetLanguage: string;
}