# lib/api-zod/src/generated/types/initiateCallResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice Translation API
 * OpenAPI spec version: 0.1.0
 */

export interface InitiateCallResponse {
  callSid: string;
  status: string;
  message?: string;
}