# lib/api-zod/src/generated/types/transcriptEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice Translation API
 * OpenAPI spec version: 0.1.0
 */

export interface TranscriptEntry {
  id: string;
  timestamp: string;
  originalText: string;
  translatedText: string;
  speaker: string;
}