# lib/api-zod/src/generated/types/translateTextRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Voice Translation API
 * OpenAPI spec version: 0.1.0
 */

export interface TranslateTextRequest {
  text: string;
  sourceLanguage: string;
  targetLanguage: string;
}